Core: Add Log class
This class work with assertion
The assertion are actived with the debug, So logging is active when debugging is active.

howto log:

assert('pixelpost\core\Log::debug("message %s %s...", $param1, $paran2, ...)');

It's important to use the full class name in assert statement unless this raise an error.

Available methods:
- pixelpost\core\Log::debug()
- pixelpost\core\Log::info()
- pixelpost\core\Log::warning()
- pixelpost\core\Log::error()