All: Add a public directory for assets and media files.
This permit to move the app folder in a non exposed internet folder.

The Plugin core class move automatically all files in the plugin's public folder to the
main public folder during the installation.
Public files are cleaned up during the plugin uninstallation.

It's no more necessary to mention the public folder in url in template.

{{ @CONTENT_URL }}admin/public/style.css
becomes:
{{ @CONTENT_URL }}admin/style.css